Extract user script infrastructure into UserScriptService
Move the JS shim template, handler registrations, and re-injection
logic from webview.dart into a dedicated UserScriptService class
in lib/services/user_script_service.dart.

WebViewFactory.createWebView() now creates a UserScriptService
instance and delegates to:
- buildInitialUserScripts() for the initialUserScripts list
- registerHandlers() in onWebViewCreated
- reinjectOnLoadStart() / reinjectOnLoadStop() for navigations

This reduces webview.dart by ~150 lines and puts all user script
concerns (shim template, URL fetching, CORS bypass, re-injection)
in one cohesive service file.
